# server/swagger/swagger.yaml
definitions:
...
  api.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  api.Meta:
    properties:
      error:
        type: string
      msg:
        description: Status int64  `json:"status"`
        type: string
    type: object
...
  api.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/login:
    post:
      description: 用户名长度不少于3位，密码不少于6位
      parameters:
      - in: formData
        maxLength: 20
        minLength: 8
        name: password
        required: true
        type: string
      - in: formData
        maxLength: 30
        minLength: 5
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户登录
      tags:
      - 用户相关
...
  /api/v1/user/delete:
    delete:
      description: 删除用户
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户IDs
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 删除用户
      tags:
      - 用户相关
  /api/v1/user/getActLog:
    get:
      description: 获取用户操作记录，不包含get
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取用户操作记录
      tags:
      - 用户相关
  /api/v1/user/getAssGroup:
    get:
      description: 返回用户关联的组
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取关联组
      tags:
      - 用户相关
  /api/v1/user/getSelfAssGroup:
    get:
      description: 返回用户本身关联的组
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取用户本身关联的组
      tags:
      - 用户相关
  /api/v1/user/getSelfInfo:
    get:
      description: 获取用户个人信息
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户个人信息
      tags:
      - 用户相关
  /api/v1/user/keyFile:
    post:
      description: '是私钥不要提交公钥！私钥如: id_rsa'
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 私钥文件上传
        in: formData
        name: keyFile
        required: true
        type: file
      - description: 私钥通行证密码上传
        in: formData
        name: Passphrase
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 提交自身私钥文件
      tags:
      - 用户相关
  /api/v1/user/keyStr:
    post:
      description: '是私钥字符串不要提交公钥文件！私钥如: id_rsa的内容'
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 私钥文本内容上传
        in: formData
        name: keyStr
        required: true
        type: string
      - description: 私钥通行证密码上传
        in: formData
        name: Passphrase
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 提交自身私钥字符串
      tags:
      - 用户相关
  /api/v1/user/logout:
    put:
      description: 登出 - 把JWT拉入黑名单
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 登出
      tags:
      - 用户相关
  /api/v1/user/password:
    patch:
      description: 修改用户密码
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: id
        required: true
        type: integer
      - description: 要求密码长度不小于8不大于20
        in: formData
        maxLength: 20
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 修改用户密码
      tags:
      - 用户相关
  /api/v1/user/search:
    get:
      description: 获取用户列表(IDs直接取用户无需其他参数，否则需要name和pageinfo)
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取用户列表
      tags:
      - 用户相关
  /api/v1/user/selfPassword:
    patch:
      description: 修改用户自己的密码
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要修改的密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 修改用户自己的密码
      tags:
      - 用户相关
  /api/v1/user/status:
    patch:
      description: status参数：恢复用户传1，禁用用户传2
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: id
        required: true
        type: integer
      - in: formData
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 修改用户状态
      tags:
      - 用户相关
  /api/v1/user/update:
    post:
      description: 新增不用传用户ID，修改才传用户ID
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: expiration
        required: true
        type: integer
      - description: 修改才需要传，没有传算新增
        in: formData
        name: id
        type: integer
      - description: 管理员传1，不然不用传
        in: formData
        name: is_admin
        type: integer
      - in: formData
        name: mobile
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        maxLength: 30
        minLength: 5
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 新增/修改用户信息
      tags:
      - 用户相关
swagger: "2.0"